Combine handlers/common and handlers/errors into handlers/responders
